Separated lib into different modules
Now consists of key, uuid, and gen_char
