+++ keysmith/src/gen_char.rs
