+++ ks-tests/src/test_key_gen.rs
@@ -1,4 +1,4 @@
-use keysmith;
+use keysmith::{key};
 
 pub struct TestKeyOpts {
     pub keys: bool,
@@ -19,7 +19,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Keys:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_key(32);
+            let key: String = key::gen_key(32);
             println!("{}", key);
         }
     }
@@ -28,7 +28,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Numbers only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_nums(32);
+            let key: String = key::gen_nums(32);
             println!("{}", key);
         }
     }
@@ -37,7 +37,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_letters(32);
+            let key: String = key::gen_letters(32);
             println!("{}", key);
         }
     }
@@ -46,7 +46,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Lowercase letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_letters_lower(32);
+            let key: String = key::gen_letters_lower(32);
             println!("{}", key);
         }
     }
@@ -55,7 +55,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Uppercase letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_letters_upper(32);
+            let key: String = key::gen_letters_upper(32);
             println!("{}", key);
         }
     }
@@ -64,7 +64,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Numbers and letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_nums_and_letters(32);
+            let key: String = key::gen_nums_and_letters(32);
             println!("{}", key);
         }
     }
@@ -73,7 +73,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Numbers and lowercase letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_nums_and_letters_lower(32);
+            let key: String = key::gen_nums_and_letters_lower(32);
             println!("{}", key);
         }
     }
@@ -82,7 +82,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Numbers and lowercase letters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_nums_and_letters_upper(32);
+            let key: String = key::gen_nums_and_letters_upper(32);
             println!("{}", key);
         }
     }
@@ -91,7 +91,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Special characters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_special_chars(32);
+            let key: String = key::gen_special_chars(32);
             println!("{}", key);
         }
     }
@@ -100,7 +100,7 @@ pub fn test_key_gen(num_of_keys: u32, opts: TestKeyOpts) {
         println!("Unsafe special characters only:");
         println!("");
         for _n in 1..=num_of_keys {
-            let key: String = keysmith::gen_special_chars_unsafe(32);
+            let key: String = key::gen_special_chars_unsafe(32);
             println!("{}", key);
         }
     }
